Extraer búsqueda manual por CI de terminal.js
Quinto paso de la descomposición de terminal.js: showManualSearchLink(),
hideManualSearchLink(), openManualSearch(), closeManualSearch(),
renderManualSearchResults() y selectManualCandidate() se mueven a
terminal/manual-search.js, junto con el cableado de sus listeners
globales (botón abrir/cancelar, click en backdrop, input de búsqueda),
reemplazado por una llamada a initManualSearch().

selectManualCandidate() mutaba terminalState.manualCandidate/
consecutiveFailures y llamaba a updateStatus() — ambos siguen siendo
responsabilidad de terminal.js, ahora pasados como callback onSelect a
initManualSearch(). Los 4 call sites externos (hideManualSearchLink/
closeManualSearch en startIdentificationFlow, hideManualSearchLink/
showManualSearchLink en identifyEmployee) no cambian.

Refactor puro, sin cambios de comportamiento.
